% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{A.frac}
\alias{A.frac}
\title{Calculate the Fraction of A-arm Selection of Two-armed Bernoulli-outcomes.}
\usage{
A.frac(x)
}
\arguments{
\item{x}{a vector of 4 elements including A$Success, A$Failure, B$Success, and B$Failure.}
}
\value{
the fraction of A-arm selected
}
\description{
Calculate the fraction of  A_armed for every 2X2 table numerated in the serialTables () when N patients to be treated.
}
\details{
A and B are examples of two arms, which resulting in favorable outcomes (success) and unfavorable outcomes (failure).
}
\examples{
AS<-13
#  the successful counts of A treatment
AF<-5
#  the failures of A treatment
BS<-2
# the successful counts of B treatment
BF<-1
#the failures of B treatment
N<- AS+AF+BS+BF
A.frac(x=c(AS,AF,BS,BF))

library(combinat) 
N <- 10
out <- serialTables(N)
ABSF<-sapply(out$Z, unlist)
A.fraction<-apply(ABSF,1,A.frac)
# the fraction of A-arm selected per table.
A.fraction
}
\keyword{NA}

