% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{cls.fun}
\alias{cls.fun}
\title{Nine Classification in Comparison of OSRs of Different Types of Decision Attitudes Popualtion.}
\usage{
cls.fun(D.mat, w)
}
\arguments{
\item{D.mat}{a matrix of difference of OSRs between homogeneous T.st and E.st with rows (the length of w), columns (a,b pairs).}

\item{w}{a vector of the degree of homogeneoudecision attitudes.}
}
\value{
a matrix with the 5050 rows (number of a,b pairs) and 3 columns with separately representing a, b, and numeric index of classification (1,2,...,9).
}
\description{
Rgeardless of decision attitudes (w values), we builded nine classifications based on comparison of OSRs between homogeneous T.st(optimistic/pessimistic attitudes) and E.st population(expected attitudes).
}
\details{
Nine classifications based on 2 w levels including -1< w <0 and 0<w<1, and 3 levels including OSRs of T.st - E.st  are all  non-negtives (+/0) or all negatives (-) with regardless of w values, or sometimes positives at particular w cases or sometimes negatives (+/-).

Nine possible classifications numerized as integers from 1 to 9.

1 states that regardless of w, (optimistic/pessimistic) T.st population perform no worser than E.st

2 states that regardless of w, the pessimist perform better or equal than E.st , however the optimist with outperforming E.st depend on w value.

3 states that regardless of w, values, the pessimist perform best, but the optimist performs worest.

4 states that regardless of w, the optimist perform best, however E.st or the pessimist either better or worser depend on the w values.

5 states that which type of population outperform depend on w values.

6 states that the optimist are worest, and who, the E.st or the pessimist,perform better depend on w values.

7 states that regardless of w, the optimist perform no worser than E.st, and E.st perform better than the pessimist.

8 states that regardless of w, the pessimist perform no worest, others depend on w values.

9 states that regardless of w, the E.st population perform T.st whatever optimistic or pessimistic attitudes

Actually, in terms of without caring w values, 1,4,7 classifications together indicate the optimist outperform E.st, 1,2,3 together indicate the pessimist outperform E.st, and 9 indicates the E.st population outperform better than T.st.

The loaded dataset in the examples is simulated based on the conditions that w = {-0.99,-0.98,...,0,0.01,...,0.98,0.99}, N= 1,2,.,100, for (a,b) = {(a,b)|a,b \in {0.01,0.02,.,0.98,0.99},a>=b}.
}
\examples{
load("mean.success.per.n.exp.100.201.RData")
load("mean.success.per.n.wpbeta.100.201.RData")

N <- 100
ps<-combn(seq(0,1,by=0.01),2)
w<-seq(from=-1,to=1,by=0.01)
Difs<-array(0,c(length(w),length(ps[1,]),N))
   for(i in 1:N){
     Difs[,,i]<-t(t(mean.success.per.n.wpbeta[,,i+1])-mean.success.per.n.exp[i+1,])
   }
  
index.100<-cls.fun(D.mat=Difs[,,N],w=w)
# When patients N=100
table(index.100[,3])  
}
\keyword{NA}

