% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{Het_func}
\alias{Het_func}
\title{The Model of Heterogeneity Decision Attitudes}
\usage{
Het_func(iter, N, a, b, u, v)
}
\arguments{
\item{iter}{the times of randomly generated from W distribution.}

\item{N}{the number of patients to be treated}

\item{a, b}{the true success rates of two arms A and B treatment}

\item{u, v}{two shape parameters of the distributions of decision attitudes w.}
}
\value{
the mean of iter times of OSR
}
\description{
Return the overall success rate (OSR) based on the model of heterogeneity decision attitudes (optimistic/pessimistic).
}
\examples{
library(combinat) 
iter<-50
N<-10
a<-0.8
b<-0.6
u<-5
v<-5
Het_func(iter,N,a,b,u,v)
}
\keyword{NA}

