% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{O_A_frac}
\alias{O_A_frac}
\title{Overall Fraction of A-arm Selection  per Individual}
\usage{
O_A_frac(fv, pv, sk)
}
\arguments{
\item{fv}{a vector of the A-arm selection fraction out of two-armed Bernoulli in all possible tables.}

\item{pv}{a vector of occurence probabilities of all tables}

\item{sk}{a list of list structure of matrices with indicating all possible Bernoulli-outcomes when N patients treated.}
}
\value{
a vector of mean of A-arm selection fraction of a series of patients n=0,1,2,...,N.
}
\description{
Average the fraction of A-arm selection per individual.
}
\examples{
library(combinat) 
N <- 10
out <- serialTables(N)
ABSF<-sapply(out$Z, unlist)
#all possible 2X2 tables

E.st_prob.A<-apply(ABSF,1,E.st_utinity)
#Probability to select the A-arm based on utinity function of E.st.
E.table.probability<-table.prob(relist(E.st_prob.A,out$sk),p=c(0.8,0.6))
E.table.probability.v<-unlist(E.table.probability)
A.fraction<-apply(ABSF,1,A.frac)
A.fraction[1]<-0.5
O_A_frac(fv=A.fraction,pv=E.table.probability.v,sk=out$sk)

#Probability to select the A-arm based on utinity function of T.st.
T.st_prob.A<-apply(ABSF,1,T.st_utinity,w=0.5)
T.table.probability<-table.prob(relist(T.st_prob.A,out$sk),p=c(0.8,0.6))
T.table.probability.v<-unlist(T.table.probability)
A.fraction<-apply(ABSF,1,A.frac)
A.fraction[1]<-0.5
O_A_frac(fv=A.fraction,pv=T.table.probability.v,sk=out$sk) 
}
\keyword{NA}

