% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{success.frac}
\alias{success.frac}
\title{Average Fraction of Favorable Outcomes of Two-armed Bernoulli-outcomes.}
\usage{
success.frac(x)
}
\arguments{
\item{x}{a vector of 4 elements inlcuding A$Success, A$Failure, B$Success, and B$Failure.}
}
\value{
average of success fraction
}
\description{
Calculate the average success fraction of two arms for every 2X2 table numerated in the serialTables () when N patients to be treated.
}
\details{
A and B are examples of two arms, which resulting in favorable outcomes (success) and unfavorable outcomes (failures).
}
\examples{
AS<-13
#  the successful counts of A treatment
AF<-5
#  the failures of A treatment
BS<-2
#  the successful counts of B treatment
BF<-1
#  the failures of B treatment
N<- AS+AF+BS+BF
success.frac(x=c(AS,AF,BS,BF))


library(combinat) 
N <- 10
out <- serialTables(N)
ABSF<-sapply(out$Z, unlist)
success.fraction<-apply(ABSF,1,success.frac)
#  the average fraction of successes per table.
success.fraction
}
\keyword{NA}

