% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{T.st_utinity}
\alias{T.st_utinity}
\title{Utinity Function Based on Targeting Decision Strategy (T.st).}
\usage{
T.st_utinity(x, w)
}
\arguments{
\item{x}{the informed informaiton of n individuals (n=0,1,2,...) with consisted of 4 integers including A$success, A$failure, B$success and B$failure counts.}

\item{w}{a degree of attitude of an individual,and is to parameterize his or her targeting value.}

\item{w}{> 0    on behalf of an optimistic individual.}

\item{w}{< 0    on behalf of a pessimistic individual.}
}
\value{
prob_A   the probability to select arm-A, which is consisted of values 1, 0.5 or 0.
}
\description{
T.st considerates the making-decision in a population with target (optimistic/pessimistic) attitudes, and calculates the probability of favorable rate more than a targeting value. 
The optimism hope their targeting values higher than the higher expected value of Beta posterior out of two. In contrast, the pessimism hope their targeting values less likely than the higher expected one.
}
\details{
A and B are examples of two treatment arms in the context, resulting into a binary outcomes success or failure after one patient to be treated. The four integeral outcomes are written as a vector.

In terms of every patient with respecting themseleves'decision attitudes, each decision process select either A or B treatment is a randomly probabilistical process which depends the output returned by their selected utinity function.

w range from -1 to 1.
}
\examples{

AS<-13
# the successful counts of A treatment
AF<-5
#the failures of A treatment
BS<-2
#the successful counts of B treatment
BF<-1
#the failures of B treatment
N<- AS+AF+BS+BF
#the total number of patients treated (n)
w=0.5
# the degree of the (n+1)-th individual's attitude who is one of the optimism. 
T.st_utinity(x=c(AS,AF,BS,BF),w=0.5)
# Decide to how much probability to select arm A for the (n+1)-th patient with an optimistic attitude. 
}
\keyword{NA}

