% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{table.prob}
\alias{table.prob}
\title{Calculate Exact Probability of a 2X2 Table with Bernoulli-outcomes}
\usage{
table.prob(s1, p)
}
\arguments{
\item{s1}{a list of probabilities to select A-arm for the all possible 2X2 tables based on E.st_utinity () or T.st_utinity ().}

\item{p}{a vector of true success rates of two arms, a and b.}
}
\value{
a list of exact probabilities per 2X2 table.
}
\description{
This function is to calculate the exact probability of every possible 2X2 table with two-armed Bernoulli-outcomes enumerated in the serialTables () when N patients to be treated.
}
\examples{
library(combinat) 
N <- 10
out <- serialTables(N)
ABSF<-sapply(out$Z, unlist)

#Probability to select the A-arm based on utinity function of E.st: E.st_utinity()
E.st_prob.A<-apply(ABSF,1,E.st_utinity)
E.table.probability<-table.prob(relist( E.st_prob.A,out$sk),p=c(0.8,0.6))
E.table.probability.v<-unlist(E.table.probability)
E.table.probability.v

#Probability to select the A-arm based on utinity function of T.st: T.st_utinity() 
T.st_prob.A<-apply(ABSF,1,T.st_utinity,w=0.5)
T.table.probability<-table.prob(relist(T.st_prob.A,out$sk),p=c(0.8,0.6))
T.table.probability.v<-unlist(T.table.probability)
T.table.probability.v
}
\keyword{NA}

