% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{Vis_image.cls}
\alias{Vis_image.cls}
\title{Visulization of Benefits of Homogeneous T.st and E.st Regardless of w}
\usage{
Vis_image.cls(N.p, index, interval)
}
\arguments{
\item{N.p}{patients number N.}

\item{index}{a matrices with 5050 raws (number of a,b  pairs) and 3 columns (a,b,and numeric classification integers, 1,2,...,9).}

\item{interval}{the interval of sequence of success rate a/b range from [0,1].}
}
\value{
trangle image with contour coded  with blue-red color.
}
\description{
Visulization of 9 classification after obtained numerical outputs from the above cls-fun(), and coded them with different colors in the triangle image.
}
\details{
The image is made in the example of this function is based on the simulated data ".RData" on the conditions that w = {-0.9995,-0.9990,...,0,0.0005,...,0.9990,0.9995}, N= 1,2,.,100, for (a,b) = {(a,b)|a,b = {0.01,0.02,.,0.98,0.99},a>=b}.

Differences of OSRs were calulated at each condition a,b pair (5050 pairs) and each w values (4001 values). And this dataset ".RData" loaded had been classified via function cls_fun().

"gdata" package used here, and color the blue indicates negatives and the red inicates positives in the trangle conditions of a, b pairs.

We saved the simulated data based on E.st and T.st with data form .RData separately in this package, and we can load two datasets to make the image.
}
\examples{
load("type.index.out.0.0005.RData", ex <- new.env())
names(ex)
N <- 100
interval<-0.01
ps<-combn(seq(0,1,by=interval),2)
w<-seq(-1,1,by=0.0005)
load("type.index.out.0.0005.RData")
Index.out<-type.index.out

table(Index.out[[N]][,3])
# 1    4    5    6    9 
# 1  828 3831  209  181 
#the table of 9 classification when patient N=100
b<-Index.out[[N]][which(Index.out[[N]][,3]==1),1]
b
a<-Index.out[[N]][which(Index.out[[N]][,3]==1),2]
a
# when true success rate a=1, and b=0, the classification 1 exists. 4 with 828 cases of a.b pairs  indicates the optimist outperformed E.st and 
# 9 with 181 pairs indicates E.st outperfomed T.st

library(gdata)
#visulized the above classifications with red, blue and gray separately.
N.id<-c(10,30,50,100)
par(mfrow=c(1,length(N.id)))
for(i in c(N.id)){
index<-Index.out[[i]]
N.p=i
Vis_image.cls(N.p=N.p,index=index,interval=interval)
}

}
\keyword{NA}

