% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{Vis_image.het}
\alias{Vis_image.het}
\title{Visulization of Benefits of heterogeneity T.st and E.st}
\usage{
Vis_image.het(het.D, U.v, interval, col.v)
}
\arguments{
\item{het.D}{a vector of the difference of OSRs between heterogenous T.st and E.st.}

\item{U.v}{the value of parameter of W distribution}

\item{interval}{the interval of sequence of success rate a/b range from [0,1].}

\item{col.v}{a vector of the color elements coded re-scaled by raw data points via color_scale().}
}
\value{
The trangle image with colored contour.
}
\description{
Visulization of difference of OSRs based on heterogeneous T.st and E.st on the condition of multiple ture success pairs, a, b.
}
\details{
".RData" on the conditions that,N= 1,2,.,100,for (a,b) = {(a,b)|a,b={0.05,0.1,.,0.90,0.95},a>=b}.

w generated from distribution function W() with various parameters u=v=5, 30, 70 or 461.
}
\examples{
N<-50
interval<-0.05
ps<-combn(seq(0,1,by=interval),2)
u=v=c(5,30,70,461)
load("mean.success.per.n.exp.100.ps210.RData")
exp.50<-mean.success.per.n.exp[N+1,]
load("het.OSRs.RData")
str(het.OSRs)
het<-sapply(het.OSRs,function(x)x[,N+1])
het.dif<-het-exp.50
cols<-color_scale(c(het.dif),neg.cols=c("blue","white"),pos.cols=c("white","red"))
par(mfrow=c(1,1),mar=c(3,3,3,3))
image(matrix(seq(0,1,length.out=length(c(het.dif))),ncol=1),col=cols[order(c(het.dif))], xlab=" ", ylab=" ",xaxt="n",yaxt="n")
axis(3,at=c(0.2),labels=c("Negatives"),cex.axis=1.4,font=4,las=1,lwd=4,tick=FALSE,line=-0.2)
axis(3,at=c(0.75),labels=c("Positives"),cex.axis=1.4,font=4,las=1,lwd=4,tick=FALSE,line=-0.2)
se<-range(c(het.dif))
axis(3,at=c(0,0.5,1),labels=round(c(min(het.dif),median(het.dif),max(het.dif)),3),cex.axis=1.1,font=4,las=1,line=-0.2,lwd=4,col.ticks="green",col="green")
colbar<-seq(0,1,length.out=length(cols))
axis(3,at=colbar[length(which(het.dif<0))],labels=c("0"),cex.axis=1.4,font=4,las=1,lwd=4,line=-0.2,col.ticks="green")
cols.mat<-matrix(cols, ncol=ncol(het.dif), nrow=nrow(het.dif))
library(gdata)
par(mfrow=c(1,length(u)))
for(i in seq_along(u)){
  het.D<-het.dif[,i]
  U.v<-u[i]
  col.v<-cols.mat[,i]
  Vis_image.het(het.D=het.D,U.v=U.v, interval=interval,col.v=col.v)
}
}
\keyword{NA}

