% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual R codes.r
\name{W}
\alias{W}
\title{Distribution of Heterogeneity of Decision Attitudes}
\usage{
W(n, u, v)
}
\arguments{
\item{u, v}{non-negative parameters of the Beta distribution.}
}
\value{
the attitude degree w randomly generated from the distribution.
}
\description{
Density function of decision attitudes with heterogeneous w is assumed from the modified beta distribution.
}
\details{
u=v indicates the symmetric shape of w distribution, indicating equal chance pessimism and optimism in a population.

The central zero indicates the homogeneity of expected attitude population. More flat the shape is, and more variance of heterogeneity of decision attitude is.
}
\examples{
u=5
v=5
n=10000
w_obs<-W(n,u,v)
par(mfrow=c(1,1))
hist(w_obs)
}
\keyword{NA}

